moving files, nav logo setting
