+++ src/views/Service/Records/Main.vue
@@ -3,9 +3,13 @@ router-view
 </template>
 
 <script setup>
-import {provide, ref } from 'vue';
+import {provide, ref, inject } from 'vue';
 provide('searchResult', ref(null));
 provide('recordToOpen', ref(null));
+
+let pageTitle = inject('pageTitle');
+let service = inject('service');
+pageTitle.value = `Service "${service.value.name}"`;
 </script>
 
 <style lang="less">
