+++ src/views/Service/Users/Main.vue
@@ -3,6 +3,9 @@ router-view
 </template>
 
 <script setup>
-import { provide, ref } from 'vue';
+import { provide, ref, inject } from 'vue';
 provide('searchResult', ref(null));
+let pageTitle = inject('pageTitle');
+let service = inject('service');
+pageTitle.value = `Service "${service.value.name}"`;
 </script>
